requirements.txt
win32gui
opencv-python
numpy
mss
pywin32
Pillow